Fix: rm circular dependency
