+++ src/utils/safe-migrations.ts
@@ -4,7 +4,7 @@ import { getSafeContractDeployment } from '@/services/contracts/deployments'
 import { sameAddress } from './addresses'
 import { getSafeToL2MigrationDeployment, getSafeMigrationDeployment } from '@safe-global/safe-deployments'
 import { type MetaTransactionData, OperationType, type SafeTransaction } from '@safe-global/safe-core-sdk-types'
-import type { ChainInfo, TransactionData } from '@safe-global/safe-gateway-typescript-sdk'
+import type { ChainInfo } from '@safe-global/safe-gateway-typescript-sdk'
 import { isValidMasterCopy } from '@/services/contracts/safeContracts'
 import { isMultiSendCalldata } from './transaction-calldata'
 import { decodeMultiSendData } from '@safe-global/protocol-kit/dist/src/utils'
@@ -93,32 +93,6 @@ export const prependSafeToL2Migration = (
   ]
 
   return __unsafe_createMultiSendTx(newTxs)
-}
-
-export const extractMigrationL2MasterCopyAddress = (txData: TransactionData): string | undefined => {
-  if (!isMultiSendCalldata(txData.hexData || '')) {
-    return undefined
-  }
-
-  const innerTxs = decodeMultiSendData(txData.hexData || '')
-  const firstInnerTx = innerTxs[0]
-  if (!firstInnerTx) {
-    return undefined
-  }
-
-  const safeToL2MigrationDeployment = getSafeToL2MigrationDeployment()
-  const safeToL2MigrationAddress = safeToL2MigrationDeployment?.defaultAddress
-  const safeToL2MigrationInterface = Safe_to_l2_migration__factory.createInterface()
-
-  if (
-    firstInnerTx.data.startsWith(safeToL2MigrationInterface.getFunction('migrateToL2').selector) &&
-    sameAddress(firstInnerTx.to, safeToL2MigrationAddress)
-  ) {
-    const callParams = safeToL2MigrationInterface.decodeFunctionData('migrateToL2', firstInnerTx.data)
-    return callParams[0]
-  }
-
-  return undefined
 }
 
 export const createUpdateMigration = (chain: ChainInfo): MetaTransactionData => {
